% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{download_pdfs}
\alias{download_pdfs}
\title{Download pdfs from googleDriveURL to pdfdir}
\usage{
download_pdfs(googleDriveURL, pdfdir)
}
\arguments{
\item{googleDriveURL}{A character of google drive folder URL}

\item{pdfdir}{A character of path to local folder}
}
\value{

}
\description{
Download pdfs from googleDriveURL to pdfdir
}
\examples{
None
}
