% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{flatten_columnX}
\alias{flatten_columnX}
\title{Flatten tb_pdf$colnameX to a character vector}
\usage{
flatten_columnX(tb_pdfs, colnameX)
}
\arguments{
\item{colnameX}{A character of column name}

\item{tb_pdf}{A list of columns}
}
\value{

}
\description{
Flatten tb_pdf$colnameX to a character vector
}
\examples{
None
}
