% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{get_dividedPDFs}
\alias{get_dividedPDFs}
\title{Get list of pdf that is divided into 5 parts}
\usage{
get_dividedPDFs(pdfdir)
}
\arguments{
\item{pdfdir}{A character of pdfs path}
}
\value{

}
\description{
Get list of pdf that is divided into 5 parts
}
\examples{
None
}
