% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{listPDFs_as_tibble}
\alias{listPDFs_as_tibble}
\title{Turn list_pdfs into a tibble}
\usage{
listPDFs_as_tibble(list_pdfs)
}
\arguments{
\item{list_pdfs}{}
}
\value{

}
\description{
Turn list_pdfs into a tibble
}
\examples{
None
}
