% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{pdf_extractPR_cat}
\alias{pdf_extractPR_cat}
\title{Extract PR in category give the dataframe of grade page}
\usage{
pdf_extractPR_cat(page_grade)
}
\arguments{
\item{page_grade}{a dataframe of the grade page}
}
\value{
a numeric vector of PR rank in category
}
\description{
Extract PR in category give the dataframe of grade page
}
\examples{
None.
}
